+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Restore external camera management support in the mobile app (#111)
+
 ### Changed
 
 ### Deprecated